Tests/LibWeb/Text/expected/css/FontFaceSet-load.txt

Load invalid font: PASS
Load CSS keyword as font: PASS
Load valid font: PASS
